decorators seperated and endpoint decorator added
